style: Home 파일의 코드를 분리 - 이전에 Home에 너무 많은 내용이 있던 것을 컴포넌트로 분리
