+++ src/pages/home/components/HomeLayout/PageContents/index.ts
@@ -1,0 +1,1 @@
+export { default as PageContents } from './PageContents';
