+++ src/pages/home/components/HomeLayout/devices/DesktopHome.tsx
