+++ src/pages/home/components/HomeLayout/devices/index.ts
