+++ src/pages/home/components/HomeLayout/index.ts
@@ -1,0 +1,1 @@
+export { default as HomeLayout } from './HomeLayout';
