+++ src/pages/home/contents/index.ts
@@ -1,3 +1,4 @@
+export * from './MainPage';
 export * from './SearchHistoryPage';
 export * from './DashboardPage';
 export * from './UploadDataPage';
